Support blob upload for calling functions
In previous versions of libmodal, we did not support inputs greater than 2 MiB. This returned an error from the server. This PR adds support for inputs larger than 2 MiB via S3 offload.

```
 FAIL  test/function.test.ts > FunctionCallLargeInput
ClientError: /modal.client.ModalClient/FunctionMap INVALID_ARGUMENT: The size of input 0 is 2.00MiB, which exceeds the maximum allowed size of 2.00MiB. Please reduce the size of the input and try again.
```

Multipart blob upload for calling functions is still not supported yet. This is currently set on the server for inputs greater than 128 MiB.